+++ test/test_daq_grafana.py
@@ -12,27 +12,6 @@ from test.util import mqtt_sensor, sleep
 logger = logging.getLogger(__name__)
 
 
-@pytest.fixture(scope="function")
-def reset_grafana():
-    """
-    Fixture to delete the Grafana datasource and dashboard.
-    """
-
-    logger.info('Grafana: Resetting artefacts')
-
-    for datasource in grafana.client.datasources.get():
-        if datasource['name'] == settings.influx_database:
-            datasource_id = datasource['id']
-            grafana.client.datasources[datasource_id].delete()
-            break
-
-    try:
-        grafana.client.dashboards.db[settings.grafana_dashboard].delete()
-    except GrafanaClientError as ex:
-        if '404' not in ex.message:
-            raise
-
-
 @pytest_twisted.inlineCallbacks
 def test_mqtt_to_grafana(machinery, create_influxdb, reset_influxdb, reset_grafana):
     """
@@ -60,7 +39,8 @@ def test_mqtt_to_grafana(machinery, create_influxdb, reset_influxdb, reset_grafa
     assert settings.influx_database in datasource_names
 
     logger.info('Grafana: Checking dashboard')
-    dashboard = grafana.client.dashboards.db[settings.grafana_dashboard].get()
+    dashboard_name = settings.grafana_dashboards[0]
+    dashboard = grafana.client.dashboards.db[dashboard_name].get()
     target = dashboard['dashboard']['rows'][0]['panels'][0]['targets'][0]
     assert target['measurement'] == settings.influx_measurement
     assert 'temperature' in target['query'] or 'humidity' in target['query']